Put all io-related code into new inputoutput directory, split up as needed
