+++ docs/export.rst
@@ -6,31 +6,31 @@ Exporting and Importing Designs
 Exporting Hardware Designs
 --------------------------
 
-.. autofunction:: pyrtl.verilog.output_to_verilog
+.. autofunction:: pyrtl.inputoutput.verilog.output_to_verilog
 
-.. autofunction:: pyrtl.inputoutput.output_to_firrtl
+.. autofunction:: pyrtl.inputoutput.firrtl.output_to_firrtl
 
 Exporting Testbenches
 ------------------------
 
-.. autofunction:: pyrtl.verilog.output_verilog_testbench
+.. autofunction:: pyrtl.inputoutput.verilog.output_verilog_testbench
 
 
 Importing Verilog
 -----------------
 
-.. autofunction:: pyrtl.inputoutput.input_from_blif
+.. autofunction:: pyrtl.inputoutput.blif.input_from_blif
 
 
 Outputting for Visualization
 ---------------------------
 
-.. autofunction:: pyrtl.inputoutput.output_to_trivialgraph
-.. autofunction:: pyrtl.inputoutput.output_to_graphviz
-.. autofunction:: pyrtl.inputoutput.graphviz_detailed_namer
-.. autofunction:: pyrtl.inputoutput.output_to_svg
-.. autofunction:: pyrtl.inputoutput.block_to_graphviz_string
-.. autofunction:: pyrtl.inputoutput.block_to_svg
-.. autofunction:: pyrtl.inputoutput.trace_to_html
-.. autofunction:: pyrtl.inputoutput.net_graph
+.. autofunction:: pyrtl.inputoutput.graphs.output_to_trivialgraph
+.. autofunction:: pyrtl.inputoutput.graphs.output_to_graphviz
+.. autofunction:: pyrtl.inputoutput.graphs.graphviz_detailed_namer
+.. autofunction:: pyrtl.inputoutput.graphs.output_to_svg
+.. autofunction:: pyrtl.inputoutput.graphs.block_to_graphviz_string
+.. autofunction:: pyrtl.inputoutput.graphs.block_to_svg
+.. autofunction:: pyrtl.inputoutput.graphs.trace_to_html
+.. autofunction:: pyrtl.inputoutput.graphs.net_graph
