+++ pyrtl/__init__.py
@@ -84,21 +84,22 @@ from .simulation import SimulationTrace
 from .compilesim import CompiledSimulation
 
 # input and output to file format routines
-from .inputoutput import input_from_blif
-from .inputoutput import output_to_trivialgraph
-from .inputoutput import graphviz_detailed_namer
-from .inputoutput import output_to_graphviz
-from .inputoutput import output_to_svg
-from .inputoutput import output_to_firrtl
-from .inputoutput import block_to_graphviz_string
-from .inputoutput import block_to_svg
-from .inputoutput import trace_to_html
-from .inputoutput import net_graph
+from .inputoutput.graphs import output_to_trivialgraph
+from .inputoutput.graphs import graphviz_detailed_namer
+from .inputoutput.graphs import output_to_graphviz
+from .inputoutput.graphs import output_to_svg
+from .inputoutput.graphs import block_to_graphviz_string
+from .inputoutput.graphs import block_to_svg
+from .inputoutput.graphs import trace_to_html
+from .inputoutput.graphs import net_graph
 
-# extraction to verilog and verilog testbench
-from .verilog import output_to_verilog
-from .verilog import OutputToVerilog
-from .verilog import output_verilog_testbench
+from .inputoutput.blif import input_from_blif
+
+from .inputoutput.firrtl import output_to_firrtl
+
+from .inputoutput.verilog import output_to_verilog
+from .inputoutput.verilog import OutputToVerilog
+from .inputoutput.verilog import output_verilog_testbench
 
 # different analysis and transform passes
 from .passes import common_subexp_elimination
