+++ pyrtl/inputoutput/blif.py
@@ -1,0 +1,370 @@
+"""
+Functions and helpers for reading BLIF files, updating the Block accordingly.
+"""
+
+from __future__ import print_function, unicode_literals
+import re
+import collections
+
+from ..pyrtlexceptions import PyrtlError
+from ..core import working_block
+from ..wire import WireVector, Input, Output, Const, Register
+from ..corecircuits import concat_list
+
+
+class Subcircuit:
+    """
+    This is a way to create and track per-module-instance wire names, so there
+    are not name clashes when we instantiate a module more than once.
+    """
+
+    def __init__(self, model, is_top=False, clk_set={'clk'}, block=None):
+        self.model = model
+        self.is_top = is_top
+        self.clk_set = clk_set
+        self.block = working_block(block)
+        self.inputs = {}
+        self.outputs = {}
+        self.wirevector_by_name = {}
+
+    def add_input(self, original_name, wire):
+        self.inputs[original_name] = wire
+        self.wirevector_by_name[original_name] = wire
+
+    def add_output(self, original_name, wire):
+        self.outputs[original_name] = wire
+        self.wirevector_by_name[original_name] = wire
+
+    def add_reg(self, original_name, wire):
+        self.wirevector_by_name[original_name] = wire
+
+    def add_clock(self, clock_name):
+        self.clk_set.add(clock_name)
+
+    def twire(self, x):
+        """ Find or make wire named x and return it. """
+        s = self.wirevector_by_name.get(x)
+        if s is None:
+            # Purposefully *not* setting its name to 'x', so we don't have name clashes
+            s = WireVector(bitwidth=1)
+            self.wirevector_by_name[x] = s
+        return s
+
+
+def input_from_blif(blif, block=None, merge_io_vectors=True, clock_name='clk', top_model=None):
+    """ Read an open BLIF file or string as input, updating the block appropriately.
+
+    :param blif: An open BLIF file to read
+    :param block: The block where the logic will be added
+    :param merge_io_vectors: If True, Input/Output wires whose names differ only
+        by a indexing subscript (e.g. 1-bit wires 'a[0]' and 'a[1]') will be combined
+        into a single Input/Output (e.g. a 2-bit wire 'a').
+    :param clock_name: The name of the clock (defaults to 'clk')
+    :param top_model: name of top-level model to instantiate; if None, defaults to first model
+        listed in the BLIF
+
+    If merge_io_vectors is True, then given 1-bit Input wires 'a[0]' and 'a[1]', these
+    wires will be combined into a single 2-bit Input wire 'a' that can be accessed
+    by name 'a' in the block. Otherwise if merge_io_vectors is False, the original 1-bit
+    wires will be Input wires of the block. This holds similarly for Outputs.
+
+    This assumes the following:
+      * The BLIF has been flattened and there is only a single module
+      * There is only one single shared clock and reset
+      * Output is generated by Yosys with formals in a particular order
+
+    It currently ignores the reset signal (which it assumes is input only to the flip flops).
+    """
+    import pyparsing
+    import six
+    from pyparsing import (Word, Literal, OneOrMore, ZeroOrMore,
+                           Suppress, Group, Keyword, Optional, oneOf)
+
+    block = working_block(block)
+
+    try:
+        blif_string = blif.read()
+    except AttributeError:
+        if isinstance(blif, six.string_types):
+            blif_string = blif
+        else:
+            raise PyrtlError('input_blif expecting either open file or string')
+
+    def SKeyword(x):
+        return Suppress(Keyword(x))
+
+    def SLiteral(x):
+        return Suppress(Literal(x))
+
+    # Begin BLIF language definition
+    signal_start = pyparsing.alphas + '$:[]_<>\\/?'
+    signal_middle = pyparsing.alphas + pyparsing.nums + '$:[]_<>\\/.?-'
+    signal_id = Word(signal_start, signal_middle)
+    header = SKeyword('.model') + signal_id('model_name')
+    input_list = Group(SKeyword('.inputs') + ZeroOrMore(signal_id))('input_list')
+    output_list = Group(SKeyword('.outputs') + ZeroOrMore(signal_id))('output_list')
+
+    cover_atom = Word('01-')
+    cover_list = Group(ZeroOrMore(cover_atom))('cover_list')
+    namesignal_list = Group(OneOrMore(signal_id))('namesignal_list')
+    name_def = Group(SKeyword('.names') + namesignal_list + cover_list)('name_def')
+
+    # asynchronous Flip-flop
+    dffas_formal = (SLiteral('C=') + signal_id('C')
+                    + SLiteral('R=') + signal_id('R')
+                    + SLiteral('D=') + signal_id('D')
+                    + SLiteral('Q=') + signal_id('Q'))
+    dffas_keyword = SKeyword('$_DFF_PN0_') | SKeyword('$_DFF_PP0_')
+    dffas_def = Group(SKeyword('.subckt') + dffas_keyword + dffas_formal)('dffas_def')
+
+    # synchronous Flip-flop
+    dffs_init_val = Optional(oneOf("0 1 2 3"), default=Literal("0"))
+    # TODO I think <type> and <control> ('re' and 'C') below are technically optional too
+    dffs_def = Group(SKeyword('.latch')
+                     + signal_id('D')
+                     + signal_id('Q')
+                     + SLiteral('re')
+                     + signal_id('C')
+                     + dffs_init_val('I'))('dffs_def')
+
+    # model reference
+    formal_actual = Group(signal_id('formal') + SLiteral('=')
+                          + signal_id('actual'))('formal_actual')
+    formal_actual_list = Group(OneOrMore(formal_actual))('formal_actual_list')
+    model_name = signal_id('model_name')
+    model_ref = Group(SKeyword('.subckt') + model_name + formal_actual_list)('model_ref')
+
+    command_def = name_def | dffas_def | dffs_def | model_ref
+    command_list = Group(OneOrMore(command_def))('command_list')
+
+    footer = SKeyword('.end')
+    model_def = Group(header + input_list + output_list + command_list + footer)
+    model_list = OneOrMore(model_def)
+    parser = model_list.ignore(pyparsing.pythonStyleComment)
+
+    # Begin actually reading and parsing the BLIF file
+    result = parser.parseString(blif_string, parseAll=True)
+    ff_clk_set = set([])
+    models = {}  # model name -> model, for subckt instantiation
+
+    def extract_inputs(subckt):
+        if subckt.is_top:
+            # NOTE: Assumes that:
+            # - Top-level inputs starting with the same prefix are part of the same wire
+            # - Indices start at 0
+            start_names = [re.sub(r'\[([0-9]+)\]$', '', x) for x in subckt.model['input_list']]
+            name_counts = collections.Counter(start_names)
+            for input_name in name_counts:
+                bitwidth = name_counts[input_name]
+                if input_name in subckt.clk_set:
+                    continue
+                elif bitwidth == 1:
+                    wire_in = Input(bitwidth=1, name=input_name, block=block)
+                    subckt.add_input(input_name, wire_in)
+                    block.add_wirevector(wire_in)
+                elif merge_io_vectors:
+                    wire_in = Input(bitwidth=bitwidth, name=input_name, block=block)
+                    for i in range(bitwidth):
+                        bit_name = input_name + '[' + str(i) + ']'
+                        bit_wire = WireVector(bitwidth=1, block=block)
+                        bit_wire <<= wire_in[i]
+                        subckt.add_input(bit_name, bit_wire)
+                else:
+                    for i in range(bitwidth):
+                        bit_name = input_name + '[' + str(i) + ']'
+                        wire_in = Input(bitwidth=1, name=bit_name, block=block)
+                        subckt.add_input(bit_name, wire_in)
+                        block.add_wirevector(wire_in)
+        else:
+            # For subckts:
+            # - Never merge input vectors
+            # - All inputs are 1-bit
+            for input_name in subckt.model['input_list']:
+                if input_name in subckt.clk_set:
+                    continue
+                wire_in = WireVector(bitwidth=1, block=block)  # Internal name prevents name clash
+                subckt.add_input(input_name, wire_in)
+                block.add_wirevector(wire_in)
+
+    def extract_outputs(subckt):
+        if subckt.is_top:
+            # NOTE: Assumes that:
+            # - Top-level outputs starting with the same prefix are part of the same wire
+            # - Indices start at 0
+            start_names = [re.sub(r'\[([0-9]+)\]$', '', x) for x in subckt.model['output_list']]
+            name_counts = collections.Counter(start_names)
+            for output_name in name_counts:
+                bitwidth = name_counts[output_name]
+                # To allow an output wire to be used as an argument (legal in BLIF),
+                # we need to create an intermediate wire, which will be used in twire()
+                # whenever the original wire is referenced. For example, given 2-bit Output 'a',
+                # every access to 'a[1]' will really be a reference to 'a[1]_i', a normal
+                # WireVector connected to 'a[1]'. A key property is that the name by
+                # which all other parts of the code refer to this wire doesn't change;
+                # the only thing that changes is what underlying wire is used.
+                if bitwidth == 1:
+                    bit_internal = WireVector(bitwidth=1, block=block)
+                    bit_out = Output(bitwidth=1, name=output_name, block=block)
+                    bit_out <<= bit_internal
+                    # NOTE this is important: redirecting user-visible name to internal wire
+                    subckt.add_output(output_name, bit_internal)
+                elif merge_io_vectors:
+                    wire_out = Output(bitwidth=bitwidth, name=output_name, block=block)
+                    bit_list = []
+                    for i in range(bitwidth):
+                        bit_name = output_name + '[' + str(i) + ']'
+                        bit_wire = WireVector(bitwidth=1, block=block)
+                        bit_list.append(bit_wire)
+                        subckt.add_output(bit_name, bit_wire)
+                    wire_out <<= concat_list(bit_list)
+                else:
+                    for i in range(bitwidth):
+                        bit_name = output_name + '[' + str(i) + ']'
+                        bit_internal = WireVector(bitwidth=1, block=block)
+                        bit_out = Output(bitwidth=1, name=bit_name, block=block)
+                        bit_out <<= bit_internal
+                        # NOTE this is important: redirecting user-visible name to internal wire
+                        subckt.add_output(bit_name, bit_internal)
+        else:
+            # For subckts:
+            # - Never merge outputs vectors
+            # - All outputs are 1-bit
+            for output_name in subckt.model['output_list']:
+                bit_out = WireVector(bitwidth=1, block=block)
+                block.add_wirevector(bit_out)
+                subckt.add_output(output_name, bit_out)
+
+    def extract_commands(subckt):
+        # for each "command" (dff or net) in the model
+        for command in subckt.model['command_list']:
+            # if it is a net (specified as a cover)
+            if command.getName() == 'name_def':
+                extract_cover(subckt, command)
+            # else if the command is a d flop flop
+            elif command.getName() == 'dffas_def' or command.getName() == 'dffs_def':
+                extract_flop(subckt, command)
+            elif command.getName() == 'model_ref':
+                extract_model_reference(subckt, command)
+            else:
+                raise PyrtlError('unknown command type')
+
+    def extract_cover(subckt, command):
+
+        def twire(w):
+            return subckt.twire(w)
+
+        # pylint: disable=invalid-unary-operand-type
+        netio = command['namesignal_list']
+        if len(command['cover_list']) == 0:
+            output_wire = twire(netio[0])
+            output_wire <<= Const(0, bitwidth=1, block=block)  # const "FALSE"
+        elif command['cover_list'].asList() == ['1']:
+            output_wire = twire(netio[0])
+            output_wire <<= Const(1, bitwidth=1, block=block)  # const "TRUE"
+        elif command['cover_list'].asList() == ['1', '1']:
+            # Populate clock list if one input is already a clock
+            if (netio[1] in subckt.clk_set):
+                subckt.add_clock(netio[0])
+            elif (netio[0] in subckt.clk_set):
+                subckt.add_clock(netio[1])
+            else:
+                output_wire = twire(netio[1])
+                output_wire <<= twire(netio[0])  # simple wire
+        elif command['cover_list'].asList() == ['0', '1']:
+            output_wire = twire(netio[1])
+            output_wire <<= ~ twire(netio[0])  # not gate
+        elif command['cover_list'].asList() == ['11', '1']:
+            output_wire = twire(netio[2])
+            output_wire <<= twire(netio[0]) & twire(netio[1])  # and gate
+        elif command['cover_list'].asList() == ['00', '1']:
+            output_wire = twire(netio[2])
+            output_wire <<= ~ (twire(netio[0]) | twire(netio[1]))  # nor gate
+        elif command['cover_list'].asList() == ['1-', '1', '-1', '1']:
+            output_wire = twire(netio[2])
+            output_wire <<= twire(netio[0]) | twire(netio[1])  # or gate
+        elif command['cover_list'].asList() == ['10', '1', '01', '1']:
+            output_wire = twire(netio[2])
+            output_wire <<= twire(netio[0]) ^ twire(netio[1])  # xor gate
+        elif command['cover_list'].asList() == ['1-0', '1', '-11', '1']:
+            output_wire = twire(netio[3])
+            output_wire <<= (twire(netio[0]) & ~ twire(netio[2])) \
+                | (twire(netio[1]) & twire(netio[2]))   # mux
+        elif command['cover_list'].asList() == ['-00', '1', '0-0', '1']:
+            output_wire = twire(netio[3])
+            output_wire <<= (~twire(netio[1]) & ~twire(netio[2])) \
+                | (~twire(netio[0]) & ~twire(netio[2]))
+        else:
+            raise PyrtlError('Blif file with unknown logic cover set "%s"'
+                             '(currently gates are hard coded)' % command['cover_list'])
+
+    def extract_flop(subckt, command):
+
+        def twire(w):
+            return subckt.twire(w)
+
+        if(command['C'] not in ff_clk_set):
+            ff_clk_set.add(command['C'])
+
+        # Create register and assign next state to D and output to Q
+        regname = command['Q'] + '_reg'
+        flop = Register(bitwidth=1)
+        subckt.add_reg(regname, flop)
+        flop.next <<= twire(command['D'])
+        flop_output = twire(command['Q'])
+        init_val = command['I']
+        if init_val == "1":
+            # e.g. in Verilog: `initial reg <= 1;`
+            raise PyrtlError("Initializing latches to 1 is not supported. "
+                             "Acceptable values are: 0, 2 (don't care), and 3 (unknown); "
+                             "in any case, PyRTL will ensure all stateful elements come up 0. "
+                             "For finer control over the initial value, use specialized reset "
+                             "logic.")
+        flop_output <<= flop
+
+    def extract_model_reference(parent, command):
+
+        def twire(w):
+            return parent.twire(w)
+
+        def get_formal_connected_to_parent_clocks():
+            clks = set()
+            for fa in command['formal_actual_list']:
+                if fa['actual'] in parent.clk_set:
+                    clks.add(fa['formal'])
+            return clks
+        formal_clks = get_formal_connected_to_parent_clocks()
+
+        subckt = Subcircuit(models[command['model_name']], clk_set=formal_clks, block=block)
+        instantiate(subckt)
+        for fa in command['formal_actual_list']:
+            formal = fa['formal']
+            actual = fa['actual']
+            if actual in parent.clk_set:
+                assert(formal in subckt.clk_set)
+                # We didn't create an input wire corresponding to this.
+                continue
+            elif formal in subckt.inputs:
+                wf = subckt.inputs[formal]
+                wa = twire(actual)
+                wf <<= wa
+            elif formal in subckt.outputs:
+                wf = subckt.outputs[formal]
+                wa = twire(actual)
+                wa <<= wf
+            else:
+                raise PyrtlError("%s formal parameter is neither an input nor output of subckt %s"
+                                 % (formal, command['model_name']))
+
+    def instantiate(subckt):
+        extract_inputs(subckt)
+        extract_outputs(subckt)
+        extract_commands(subckt)
+
+    # Get all model definitions
+    for model in result:
+        if not top_model:
+            top_model = model['model_name']
+        models[model['model_name']] = model
+
+    top = Subcircuit(models[top_model], is_top=True, clk_set={clock_name}, block=block)
+    instantiate(top)
