+++ tests/inputoutput/test_blif.py
@@ -1,12 +1,7 @@
 import unittest
-import random
-import io
 import pyrtl
 import six
-from pyrtl import inputoutput
-from pyrtl import analysis
 from pyrtl.rtllib import testingutils as utils
-
 
 full_adder_blif = """\
 # Generated by Yosys 0.3.0+ (git sha1 7e758d5, clang 3.4-1ubuntu3 -fPIC -Os)
@@ -643,257 +638,6 @@ class TestInputFromBlif(unittest.TestCase):
         })
         cvals = sim.tracer.trace[c]
         self.assertEqual(cvals, [0, 0, 0, 1, 1, 1, 0])
-
-
-graphviz_string = """
-              digraph g {
-
-              graph [splines="spline"];
-              node [shape=circle, style=filled, fillcolor=lightblue1,
-                    fontcolor=black, fontname=helvetica, penwidth=0,
-                    fixedsize=true];
-              edge [labelfloat=false, penwidth=2, color=deepskyblue, arrowsize=.5];
-                  n0 [label="bits(0,0) (Fanout: 1)", height=.1, width=.1];
-    n1 [label="bits(7,2) (Fanout: 1)", height=.1, width=.1];
-    n2 [label="bits(0,0) (Fanout: 1)", height=.1, width=.1];
-    n3 [label="concat (Fanout: 1)", height=.1, width=.1];
-    n4 [label=" (Fanout: 0)", height=.1, width=.1];
-    n5 [label="concat (Fanout: 1)", height=.1, width=.1];
-    n6 [label="* (Fanout: 1)"];
-    n7 [label="8", shape=circle, fillcolor=lightgrey];
-    n8 [label="d", shape=house, fillcolor=lawngreen];
-    n9 [label="0", shape=circle, fillcolor=lightgrey];
-    n10 [label="0", shape=circle, fillcolor=lightgrey];
-    n11 [label="a", shape=invhouse, fillcolor=coral];
-   n0 -> n3 [label="tmp0/2 (Delay: 0.00)", penwidth="6", arrowhead="none"];
-   n1 -> n5 [label="tmp3/6 (Delay: 706.50)", penwidth="6", arrowhead="none"];
-   n2 -> n5 [label="tmp4/4 (Delay: 0.00)", penwidth="6", arrowhead="none"];
-   n3 -> n6 [label="tmp1/4 (Delay: 0.00)", penwidth="6", arrowhead="normal"];
-   n4 -> n8 [label="d/10 (Delay: 706.50)", penwidth="6", arrowhead="normal"];
-   n5 -> n4 [label="tmp5/10 (Delay: 706.50)", penwidth="6", arrowhead="normal"];
-   n6 -> n1 [label="tmp2/8 (Delay: 706.50)", penwidth="6", arrowhead="none"];
-   n7 -> n6 [label="const_0_8/4 (Delay: 0.00)", penwidth="6", arrowhead="normal"];
-   n9 -> n0 [label="const_1_0/1 (Delay: 0.00)", penwidth="2", arrowhead="none"];
-   n10 -> n2 [label="const_2_0/1 (Delay: 0.00)", penwidth="2", arrowhead="none"];
-   n11 -> n3 [label="a/2 (Delay: 0.00)", penwidth="6", arrowhead="none"];
-}
-
-
-"""
-
-
-class TestOutputGraphs(unittest.TestCase):
-    def setUp(self):
-        pyrtl.reset_working_block()
-
-    def test_output_to_tgf_does_not_throw_error(self):
-        with io.StringIO() as vfile:
-            pyrtl.input_from_blif(full_adder_blif)
-            pyrtl.output_to_trivialgraph(vfile)
-
-    def test_output_to_graphviz_does_not_throw_error(self):
-        with io.StringIO() as vfile:
-            pyrtl.input_from_blif(full_adder_blif)
-            pyrtl.output_to_graphviz(vfile)
-
-    def test_output_to_graphviz_with_custom_namer_does_not_throw_error(self):
-        with io.StringIO() as vfile:
-            pyrtl.input_from_blif(full_adder_blif)
-            timing = analysis.TimingAnalysis()
-            node_fan_in = {net: len(net.args) for net in pyrtl.working_block()}
-            graph_namer = pyrtl.inputoutput.graphviz_detailed_namer(
-                extra_node_info=node_fan_in,
-                extra_edge_info=timing.timing_map
-            )
-            pyrtl.output_to_graphviz(vfile, namer=graph_namer)
-
-    @unittest.skip("Need to make Graphviz output order deterministic via sorting")
-    def test_output_to_graphviz_correct_detailed_output(self):
-        from pyrtl.analysis.estimate import TimingAnalysis
-
-        a = pyrtl.Input(2, 'a')
-        b = a * 8
-        c = b[2:]
-        d = pyrtl.Output(10, 'd')
-        d <<= c
-
-        analysis = TimingAnalysis()
-        _, dst_map = pyrtl.working_block().net_connections()
-
-        def get_fanout(n):
-            if isinstance(n, pyrtl.LogicNet):
-                if n.op == '@':
-                    return 0
-                w = n.dests[0]
-            else:
-                w = n
-
-            if isinstance(w, pyrtl.Output):
-                return 0
-            else:
-                return len(dst_map[w])
-
-        node_fanout = {n: "Fanout: %d" % get_fanout(n) for n in pyrtl.working_block().logic}
-        wire_delay = {
-            w: "Delay: %.2f" % analysis.timing_map[w] for w in pyrtl.working_block().wirevector_set
-        }
-
-        with io.StringIO() as vfile:
-            pyrtl.output_to_graphviz(
-                file=vfile,
-                namer=pyrtl.graphviz_detailed_namer(node_fanout, wire_delay)
-            )
-            self.assertEqual(vfile.getvalue(), graphviz_string)
-
-
-class TestNetGraph(unittest.TestCase):
-    def setUp(self):
-        pyrtl.reset_working_block()
-
-    def test_as_graph(self):
-        inwire = pyrtl.Input(bitwidth=1, name="inwire1")
-        inwire2 = pyrtl.Input(bitwidth=1)
-        inwire3 = pyrtl.Input(bitwidth=1)
-        tempwire = pyrtl.WireVector()
-        tempwire2 = pyrtl.WireVector()
-        outwire = pyrtl.Output()
-
-        tempwire <<= inwire | inwire2
-        tempwire2 <<= ~tempwire
-        outwire <<= tempwire2 & inwire3
-
-        g = inputoutput.net_graph()
-        # note for future: this might fail if we change
-        # the way that temp wires are inserted, but that
-        # should not matter for this test and so the number
-        # can be safely updated.
-        self.assertEqual(len(g), 10)
-
-    def test_netgraph_unused_wires(self):
-        genwire = pyrtl.WireVector(8, "genwire")
-        inwire = pyrtl.Input(8, "inwire")
-        outwire = pyrtl.Output(8, "outwire")
-        constwire = pyrtl.Const(8, 8)
-        reg = pyrtl.Register(8, "reg")
-        g = inputoutput.net_graph()
-        self.assertEquals(len(g), 0)
-
-
-class TestOutputIPynb(unittest.TestCase):
-    def setUp(self):
-        pyrtl.reset_working_block()
-        self.maxDiff = None
-
-    def test_one_bit_adder_matches_expected(self):
-        temp1 = pyrtl.WireVector(bitwidth=1, name='temp1')
-        temp2 = pyrtl.WireVector()
-
-        a, b, c = pyrtl.Input(1, 'a'), pyrtl.Input(1, 'b'), pyrtl.Input(1, 'c')
-        sum, carry_out = pyrtl.Output(1, 'sum'), pyrtl.Output(1, 'carry_out')
-
-        sum <<= a ^ b ^ c
-
-        temp1 <<= a & b  # connect the result of a & b to the pre-allocated wirevector
-        temp2 <<= a & c
-        temp3 = b & c  # temp3 IS the result of b & c (this is the first mention of temp3)
-        carry_out <<= temp1 | temp2 | temp3
-
-        sim_trace = pyrtl.SimulationTrace()
-        sim = pyrtl.Simulation(tracer=sim_trace)
-        for cycle in range(15):
-            sim.step({
-                'a': random.choice([0, 1]),
-                'b': random.choice([0, 1]),
-                'c': random.choice([0, 1])
-            })
-
-        htmlstring = inputoutput.trace_to_html(sim_trace)  # tests if it compiles or not
-
-
-firrtl_output_concat_test = """\
-circuit Example :
-  module Example :
-    input clock : Clock
-    input reset : UInt<1>
-    output o : UInt<13>
-    wire tmp0 : UInt<13>
-    wire tmp1 : UInt<7>
-    wire tmp2 : UInt<13>
-    node const_0_12 = UInt<4>(12)
-    node const_1_3 = UInt<3>(3)
-    node const_2_38 = UInt<6>(38)
-
-    o <= tmp0
-    tmp0 <= tmp2
-    tmp1 <= cat(const_0_12, const_1_3)
-    tmp2 <= cat(tmp1, const_2_38)
-"""
-
-firrtl_output_select_test = """\
-circuit Example :
-  module Example :
-    input clock : Clock
-    input reset : UInt<1>
-    output b : UInt<6>
-    wire tmp0 : UInt<6>
-    wire tmp1 : UInt<1>
-    wire tmp2 : UInt<1>
-    wire tmp3 : UInt<1>
-    wire tmp4 : UInt<1>
-    wire tmp5 : UInt<1>
-    wire tmp6 : UInt<1>
-    wire tmp7 : UInt<6>
-    wire tmp8 : UInt<2>
-    wire tmp9 : UInt<3>
-    wire tmp10 : UInt<4>
-    wire tmp11 : UInt<5>
-    wire tmp12 : UInt<6>
-    node const_0_2893 = UInt<12>(2893)
-
-    b <= tmp0
-    tmp0 <= tmp7
-    tmp1 <= bits(const_0_2893, 0, 0)
-    tmp2 <= bits(const_0_2893, 2, 2)
-    tmp3 <= bits(const_0_2893, 4, 4)
-    tmp4 <= bits(const_0_2893, 6, 6)
-    tmp5 <= bits(const_0_2893, 8, 8)
-    tmp6 <= bits(const_0_2893, 10, 10)
-    tmp7 <= tmp12
-    tmp8 <= cat(tmp6, tmp5)
-    tmp9 <= cat(tmp8, tmp4)
-    tmp10 <= cat(tmp9, tmp3)
-    tmp11 <= cat(tmp10, tmp2)
-    tmp12 <= cat(tmp11, tmp1)
-"""
-
-
-class TestOutputFirrtl(unittest.TestCase):
-    def setUp(self):
-        pyrtl.reset_working_block()
-        pyrtl.wire._reset_wire_indexers()
-        pyrtl.memory._reset_memory_indexer()
-
-    def test_textual_consistency_concats(self):
-        i = pyrtl.Const(0b1100)
-        j = pyrtl.Const(0b011, bitwidth=3)
-        k = pyrtl.Const(0b100110)
-        o = pyrtl.Output(13, 'o')
-        o <<= pyrtl.concat(i, j, k)
-
-        buffer = io.StringIO()
-        pyrtl.output_to_firrtl(buffer)
-
-        self.assertEqual(buffer.getvalue(), firrtl_output_concat_test)
-
-    def test_textual_consistency_selects(self):
-        a = pyrtl.Const(0b101101001101)
-        b = pyrtl.Output(6, 'b')
-        b <<= a[::2]
-
-        buffer = io.StringIO()
-        pyrtl.output_to_firrtl(buffer)
-
-        self.assertEqual(buffer.getvalue(), firrtl_output_select_test)
 
 
 if __name__ == "__main__":
